+++ src/model.cpp
@@ -1,6 +1,7 @@
 #include "model.h"
 
 #include <algorithm>
+#include <functional>
 
 #include "action.h"
 #include "controller.h"
@@ -295,6 +296,12 @@ void Model::toggle_case() {
             }
         } break;
 
+        case ActionType::DelCurrentWord: {
+            if (cur_change.text.has_value()) {
+                buf.at(current_line).insert(current_char, cur_change.text.value());
+            }
+        } break;
+
         default:
             break;
     };
@@ -367,6 +374,10 @@ void Model::toggle_case() {
             delete_current_line();
         } break;
 
+        case ActionType::DelCurrentWord: {
+            delete_current_word({cur_change.text.value(), current_char, current_line});
+        } break;
+
         default:
             break;
     };
@@ -404,4 +415,33 @@ void Model::delete_current_line() {
     }
 
     unsaved = true;
+}
+
+[[nodiscard]] const WordPos Model::current_word() const {
+    WordPos ret = {"", 0, 0};
+    const std::string* cur_line = &buf.at(current_line);
+    uint_t start = current_char;
+
+    while (start && is_letter(cur_line->at(start))) {
+        start--;
+    }
+
+    // if we aren't at 0, we're on a whitespace, so we want to increment once
+    if (start) {
+        start++;
+    }
+
+    uint_t len = uint_t(
+        std::distance(
+            cur_line->begin() + start,
+            std::find_if(cur_line->begin() + start, cur_line->end(), std::not_fn(is_letter))));
+
+    ret.start_pos = start;
+    ret.text = cur_line->substr(start, len);
+    ret.lineno = current_line;
+    return ret;
+}
+
+void Model::delete_current_word(const WordPos pos) {
+    buf.at(pos.lineno).erase(pos.start_pos, pos.text.size());
 }